llm-chatbot-main/llm-chatbot-main/requirements.txt
streamlit
langchain
langchain_community
sentence-transformers
faiss-cpu
llama-cpp-python
numpy
asyncio
pypdf